requirements.txt
Flask
openpyxl
pdfplumber
pytesseract
Pillow
